requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.19
Jinja2==3.1.4
attrs==24.2.0
slowapi==0.1.9
SQLAlchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0
asyncmy==0.2.10
aiomysql==0.2.0
aiopg==1.4.0
asyncpg==0.30.0
psycopg2==2.9.10
pydantic-settings==2.6.1
python-decouple==3.8
